Add Base Classes for Entities

+++ TowerForge/TowerForge/LevelManager/Components/HealthComponent.swift
@@ -8,20 +8,20 @@
 import Foundation
 
 class HealthComponent: TFComponent {
-    public var currentHealth: Int
-    public var maxHealth: Int
-    
-    init(maxHealth: Int) {
+    public var currentHealth: Float
+    public var maxHealth: Float
+
+    init(maxHealth: Float) {
         self.currentHealth = maxHealth
         self.maxHealth = maxHealth
         super.init()
     }
     
-    func decreaseHealth(amount: Int) {
+    func decreaseHealth(amount: Float) {
         self.currentHealth -= amount
     }
     
-    func increaseHealth(amount: Int) {
+    func increaseHealth(amount: Float) {
         self.currentHealth = min(self.currentHealth + amount, self.maxHealth)
     }
